% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_binary.R
\name{add.missing.vars_xgb}
\alias{add.missing.vars_xgb}
\title{add.missing.vars_xgb}
\usage{
add.missing.vars_xgb(xgb.model, testSet)
}
\arguments{
\item{xgb.model}{XGBoost model}

\item{testSet}{Data frame containing a CRE test set}
}
\description{
Function to add columns with zero counts for motifs present in the model but not in the test set
}
