% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_CREs.R
\name{adjust_CREs}
\alias{adjust_CREs}
\title{adjust_CREs
Function to trim (or extend) CREs}
\usage{
adjust_CREs(x, N, chrom_sizes)
}
\arguments{
\item{x}{data frame of genomic coordinates. BED file format.}

\item{N}{Number of base pairs - used to extend coordinates stored in x.}

\item{chom_sizes}{dataframe of chromosome sizes}
}
\description{
adjust_CREs
Function to trim (or extend) CREs
}
