% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOM_shap.R
\name{annonTATE}
\alias{annonTATE}
\title{Re-annotates a ggplot created from shapviz}
\usage{
annonTATE(
  plotDat,
  peakAnnotations,
  annotLength = 30,
  makeUnique = FALSE,
  waterfallOther = NULL
)
}
\arguments{
\item{plotDat}{Data slot from a ggplot object}

\item{peakAnnotations}{}

\item{annotLength}{Number of characters to truncate annotation to. Default 30 character.}

\item{makeUnique}{Make sure every MOTIF converts to a unique entry. Default FALSE. When false multiple TF entries are collapsed}

\item{waterfallOther}{Annotation for waterfall plots "other" category}
}
\description{
Re-annotates a ggplot created from shapviz
}
