% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_CREs.R
\name{filterCREs}
\alias{filterCREs}
\title{filterCREs}
\usage{
filterCREs(
  input_bed = NULL,
  annot = NULL,
  chr_sizes = NULL,
  u = NULL,
  d = NULL,
  nbp = NULL,
  keep_proximal = FALSE,
  remove_proximal = FALSE,
  non_exonic = FALSE,
  out_bed = NULL
)
}
\arguments{
\item{u}{number of basepairs upstream of transcriptional start sites}

\item{d}{number of basepairs downstream of transcriptional start site}

\item{nbp}{number of base pairs}

\item{keep_proximal}{boolean. Only keep proximal regions (defined by u and d) relative to transcription start sites}

\item{remove_proximal}{Boolean: Remove proximal regions (defined by u and d) relative to transcriptional start sites}

\item{non_exonic}{Boolean - defines if non exonic regions should ONLY be considered.}

\item{out_bed}{name of output bed file}

\item{inputBedFile}{input bed filename}

\item{annotFile}{annotation file name}
}
\description{
Function to filter Cis Regulatory Elements (CRE)
}
\examples{
\dontrun{
extdata_path <- system.file("extdata",package = "BagOfMotifs")
input.bed <- paste0(extdata_path, "/tutorial/mouseE8.25_peaks.bed")
annot.file <- paste0(extdata_path,"/Mus_musculus.GRCm38.92.gtf.gz")
mouseChrSizes <- paste0(extdata_path,"/mouse_sizes_primary_genome_tab.txt")


 filterCRE(inputBedFile = input.bed, annotFile = annot.file, 
                    remove_proximal = TRUE,  non_exonic = TRUE,
					   u =  1000, d=1000, nbp=500,
                    chrSizes = mouseChrSizes, 
                    out_bed =  'mouseE8.25_peaks_filt.bed')
}
}
