% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFasta.R
\name{generateAllFasta}
\alias{generateAllFasta}
\title{Wrapper for getFasta function. Will generate fasta files from bed files located within a specified directory}
\usage{
generateAllFasta(bedDir = NULL, genome = NULL, fastaDir = NULL)
}
\arguments{
\item{bedDir}{Path to directory containing celltype/state specific BED files. The first three columns in the BED files should correspond to chromosome, start and end positions. The fourth column should contain cell type/state annotation.}

\item{genome}{Genome as a BSgenome object}

\item{fastaDir}{Directory to where fasta files are written to}
}
\description{
Wrapper for getFasta function. Will generate fasta files from bed files located within a specified directory
}
\examples{

extdata_path <- system.file("extdata",package = "BagOfMotifs")
bed_path <- paste0(extdata_path, "/tutorial/bed_files/")
fasta_path <- paste0(extdata_path, "/tutorial/fasta_files/")

# Load genome
library("BSgenome.Mmusculus.UCSC.mm10")
Mmusculus <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10

generateAllFasta(bedDir = bed_path, genome = Mmusculus, fastaDir = fasta_path)
}
