% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFasta.R
\name{getFasta}
\alias{getFasta}
\title{Function to extract sequences for a set of CRE coordinates in BED format}
\usage{
getFasta(bed = NULL, genome, FastaFile = NULL)
}
\arguments{
\item{bed}{Data frame containing CRE coordinates. The first three columns should correspond to chromosome, start and end positions. The fourth column should contain cell type/condition annotation.}

\item{genome}{Genome as a BSgenome object}

\item{FastaFile}{Path to output fasta file. If NULL, a Biostrings object will be returned. Default = NULL.}
}
\description{
Function to extract sequences for a set of CRE coordinates in BED format
}
\examples{

extdata_path <- system.file("extdata",package = "BagOfMotifs")
bed_path <- paste0(extdata_path, "/tutorial/bed_files/Cardiomyocytes.bed")
bed_cardiom <- read.table(file = bed_path, header = F, stringsAsFactors = F
, sep = '\t')  

# Match chromosome notation
bed$V1 <- paste0("chr", bed$V1)

# Load genome
library("BSgenome.Mmusculus.UCSC.mm10")
Mmusculus <- BSgenome.Mmusculus.UCSC.mm10::BSgenome.Mmusculus.UCSC.mm10
getFasta(bed = bed_cardiom, genome = Mmusculus, FastaFile = "Cardiomyocytes.fa")
}
