% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_stats.R
\name{MCC}
\alias{MCC}
\title{Function to calculate Matthew's correlation coefficient (MCC)}
\usage{
MCC(TP, TN, FP, FN)
}
\arguments{
\item{TP}{Number of True Positive instances}

\item{TN}{Number of True Negative instances}

\item{FP}{Number of False Positive instances}

\item{FN}{Number of False Negative instances}
}
\description{
Function to calculate Matthew's correlation coefficient (MCC)
}
