% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_binary.R
\name{prcurve_binary}
\alias{prcurve_binary}
\title{prcurve_binary}
\usage{
prcurve_binary(motifs, xgb_model, title = NULL)
}
\arguments{
\item{motifs}{File containing motif counts}

\item{xgb_model}{File name of model}

\item{title}{Title for output plot (default NULL)}
}
\description{
Function to produce a precision/recall curve for binary prediction
}
\examples{
\dontrun{

extdata_path <- system.file("extdata",package = "BagOfMotifs")
motif_counts <- paste0(extdata_path, "/tutorial/motifs/Cardiomyocytes_vs_other_counts.txt")


predict_binary(motifs = motif_counts, xgb_model = paste0(extdata_path, "/tutorial/motifs/Cardiomyocytes_vs_other.rds"))
}
}
