% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_binary.R
\name{prcurve_multi}
\alias{prcurve_multi}
\title{prcurve_multi}
\usage{
prcurve_multi(
  inputMotif_dir = NULL,
  inputXGB_dir = NULL,
  ncol = 3,
  nrow = 3,
  outputFile = NULL,
  width = 12,
  height = 12
)
}
\arguments{
\item{inputMotif_dir}{Directory of motif counts (named <celltype>_vs_Others.txt)}

\item{inputXGB_dir}{Directory of XGBoost files (names <celltype>_vs_Others.rds)}

\item{ncol}{Number of columns of plots per page}

\item{nrow}{Number of rows of plots per page}

\item{outputFile}{Output pdf file name for plots}

\item{width}{Width of the graphics region in pdf output in inches (default 12)}

\item{height}{Height of the graphics region in pdf output in inches (default 12)}
}
\description{
Wrapper function for prcurve_binary. Will run 
on multiple files within specified directory
}
\examples{
\dontrun{
extdata_path <- system.file("extdata",package = "BagOfMotifs")
motif_counts <- paste0(extdata_path, "/tutorial/motifs/Cardiomyocytes_vs_other_counts.txt")


predict_binary(motifs = motif_counts, xgb_model = paste0(extdata_path, "/tutorial/motifs/Cardiomyocytes_vs_other.rds")
}

}
