% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_binary.R
\name{predict_binary_multi}
\alias{predict_binary_multi}
\title{predict_binary_multi}
\usage{
predict_binary_multi(
  inputMotif_dir = NULL,
  inputXGB_dir = NULL,
  outputTrain_dir = NULL,
  pred_dir = NULL,
  ncol = 3,
  nrow = 3,
  outputFile = NULL,
  width = 12,
  height = 12
)
}
\arguments{
\item{inputMotif_dir}{Directory of motif counts (named <celltype>_vs_Others.txt)}

\item{inputXGB_dir}{Directory of XGBoost files (names <celltype>_vs_Others.rds)}

\item{outputTrain_dir}{Directory where to save training text output files}

\item{ncol}{Number of columns of plots per page}

\item{nrow}{Number of rows of plots per page}

\item{outputFile}{Output pdf file name for plots}
}
\description{
Wrapper function for predict_binary. Will run 
on multiple files within specified directory
}
\examples{
\dontrun{
extdata_path <- system.file("extdata",package = "BagOfMotifs")
motif_counts <- paste0(extdata_path, "/tutorial/motifs/Cardiomyocytes_vs_other_counts.txt")


predict_binary(motifs = motif_counts, xgb_model = paste0(extdata_path, "/tutorial/motifs/Cardiomyocytes_vs_other.rds")
}

}
