% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_motifs.R
\name{rank_shap}
\alias{rank_shap}
\title{rankMotifs}
\usage{
rank_shap(shap_file, type, out_file)
}
\arguments{
\item{shap_file}{Path to the file containing SHAP scores}

\item{out_file}{File name to output the motifs ranked by SHAP}

\item{rank_type}{Rank type. Either 'sum' or 'mean'}
}
\description{
Function to rank motifs by the sum or mean of absolute SHAP scores.
}
\examples{

\dontrun{
extdata_path <- system.file("extdata",package = "BagOfMotifs")
binPredictions <- paste0(extdata_path, "/tutorial/[SHAP SCORES FILE]")


rankMotifs(shap_file = binPredictions, out_file = "ranked_motifs", rank_type = "sum")
}
}
