% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_motifs.R
\name{rankMotifs}
\alias{rankMotifs}
\title{rankMotifs}
\usage{
rankMotifs(shap_file, type, out_file)
}
\arguments{
\item{shap_file}{Path to the file containing SHAP scores}

\item{out_file}{File name to output the motifs ranked by SHAP (default to "ranked_motifs")}

\item{rank_type}{Rank type. Either 'sum' or 'mean'}
}
\description{
Function to rank motifs by the sum or mean of absolute SHAP scores.
}
\examples{

\dontrun{
extdata_path <- system.file("extdata",package = "BagOfMotifs")
binPredictions <- paste0(extdata_path, "/tutorial/Cardiomyocytes_vs_other_SHAP.txt")


rankMotifs(shap_file = binPredictions, out_file = "ranked_motifs", type = "sum")
}
}
