% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFIMO.R
\name{runFIMO}
\alias{runFIMO}
\title{Function to execute FIMO for at least two sets of CREs specific to different cell types/conditions}
\usage{
runFIMO(
  input_path,
  motifs_path,
  out_path,
  p_thresh = 1e-04,
  FIMO_path,
  verbosity = 1
)
}
\arguments{
\item{input_path}{Path to folder containing fasta (.fa) files with CRE sequences}

\item{motifs_path}{Motif database file (meme format)}

\item{out_path}{Output file folder. The output folder will be created and FIMO output for each fasta file will be saved to this directory.}

\item{p_thresh}{p-value threshold (default 0.0001)}

\item{FIMO_path}{Path to FIMO}

\item{verbosity}{Level of verbosity for FIMO. Default = 1.}
}
\description{
Function to execute FIMO for at least two sets of CREs specific to different cell types/conditions
}
\examples{

extdata_path <- system.file("extdata",package = "BagOfMotifs")
bed_path <- paste0(extdata_path, "/tutorial/bed_files")

# Replace /path/to/fimo with the path to FIMO tool
runFIMO(input_path = , motifs_path = paste0(extdata_path, "/gimme.vertebrate.v5.0.meme")
, out_path = "mouseE8.25_motifs", p_thresh = 0.0001
, FIMO_path = '/path/to/fimo')
}
