% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOM_shap.R
\name{save_shap}
\alias{save_shap}
\title{Saves SHAP values to a file}
\usage{
save_shap(xgb_model = NULL, ts = NULL, shap_file)
}
\arguments{
\item{xgb_model}{xgb_model prepared from xgboost}

\item{ts}{training set (matrix)}

\item{shap_file}{filename to save shap values to. Will create a tab delimited text file}
}
\description{
Saves SHAP values to a file
}
