% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOM_shap.R
\name{save_shap_multi}
\alias{save_shap_multi}
\title{Saves SHAP values to a file}
\usage{
save_shap_multi(
  dataDir = NULL,
  xgb_models = NULL,
  train_sets = NULL,
  outDir = NULL
)
}
\description{
Saves SHAP values to a file
}
