% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOM_shap.R
\name{ShapBarPlot}
\alias{ShapBarPlot}
\title{Plots a bar plot of SHAP values}
\usage{
ShapBarPlot(scores, peakAnnotations = NULL, numTF = 10, annotLength = 30)
}
\arguments{
\item{scores}{A data frame SHAP scores for a particular locus. Colnames must contain motif ID and
match entries inthe Motif column of peakAnnotations data frame.}

\item{peakAnnotations}{Dataframe of transcription factor lookup ID information. Must have columns
names Motif and Factor. Values in Motif column should match column names in 
the scores dataframe input. If NULL no annotation is provided.}

\item{numTF}{Number of transcription factors to plot. The remaining transcriptions factors are plotted under other}

\item{annotLength}{Integer indicating length of annotation. Default is 30 characters.}
}
\description{
Plots a bar plot of SHAP values
}
