% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOM_shap.R
\name{shapPlots_multi}
\alias{shapPlots_multi}
\title{Makes SHAP plots for multiple models}
\usage{
shapPlots_multi(
  dataDir = NULL,
  xgb_models = NULL,
  train_sets = NULL,
  plotType = "bar",
  max_display = 15,
  CRE_ids = NULL,
  annotDat = NULL,
  annotLength = 30,
  order = "decreasing",
  show_numbers = FALSE,
  average_shap = TRUE,
  ...
)
}
\arguments{
\item{xgb_models}{List of file names for models}

\item{train_sets}{List of file names for training sets (same order as models).}

\item{plotType}{Plottype, can be "bar" (default), "beeswarm", or "waterfall"}

\item{max_display}{Number of transcription factors to plot (Default 15). The remaining transcriptions factors are plotted under other}

\item{order}{Ïf set to "decreasing" then plots the highest value first through to lowest value.}

\item{show_numbers}{Boolean. If set to TRUE will display numbers on plot.}

\item{average_shap}{Boolean. Whether to average shap values across the specified CREs in waterfall plots (default TRUE).}

\item{annotData}{Dataframe with columns "Motif" and "Factor". The values in "Motif" should match what was used}
}
\description{
Makes SHAP plots for multiple models
}
