% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_CREs.R
\name{textToBED}
\alias{textToBED}
\title{textToBED}
\usage{
textToBED(
  inputTextFile = NULL,
  header = TRUE,
  inputcolnames = c("peak_chr", "peak_start", "peak_end", "celltype_specificity"),
  sep = ",",
  outputFileName = "out.bed",
  removeDuplicatePeaks = TRUE,
  removeMultiAnnotatedPeaks = TRUE,
  removeUnnanotatedPeaks = TRUE
)
}
\arguments{
\item{inputtextFile}{input bed filename}

\item{annotFile}{annotation file name}

\item{u}{number of basepairs upstream of transcriptional start sites}

\item{d}{number of basepairs downstream of transcriptional start site}

\item{nbp}{number of base pairs}

\item{chrSizesFile}{File name of chromsome sizes ( tab 
delimited file with no header: first column chromosome ID, second column size)}

\item{keep_proximal}{boolean. Only keep proximal regions (defined by u and d) relative to transcription start sites}

\item{remove_proximal}{Boolean: Remove proximal regions (defined by u and d) relative to transcriptional start sites}

\item{non_exonic}{Boolean - defines if non exonic regions should ONLY be considered.}

\item{out_bed}{filename of output bed file after filtering.}

\item{inputBedZeroBased}{Boolean (default TRUE). Add 1 to start position of input bed file.}

\item{celloutputDir}{directory name (will create) where individual BED files will be populated. If NULL this step is ommited.}

\item{minCellPercent}{Each cell population must make up this percent of all cells (default 1 percent). Will notify what cells are removed}
}
\description{
converts a text file to BED format for peak data
}
\examples{
\dontrun{
extdata_path <- system.file("extdata",package = "BagOfMotifs")
filename <- paste0(extdata_path,'/Pijuan_etal_table_S6.csv.gz')


 textToBed(inputTextFile = filename, outputFileName =  'mouseE8.25_peaks_filt.bed')
}

}
