% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_binary.R
\name{train_binary}
\alias{train_binary}
\title{train_binary}
\usage{
train_binary(
  input_data = NULL,
  nrounds = 10000,
  eta = 0.01,
  max_depth = 6,
  subsample = 0.5,
  colsample_bytree = 0.5,
  objective = "binary:logistic",
  early_stopping_rounds = NULL,
  nthread = 1,
  eval_metric = "error",
  maximize = F,
  params = list(),
  feval = NULL,
  verbose = 1,
  print_every_n = 1L,
  save_period = NULL,
  save_name = "xgboost.model",
  xgb_model = NULL,
  callbacks = list()
)
}
\arguments{
\item{input_data}{File containing the input matrix of motif counts}

\item{nrounds}{Number of boosting rounds (default: 10000)}

\item{eta}{Learning rate (default: 0.01)}

\item{max_depth}{Maximum tree depth (default: 6)}

\item{subsample}{Subsample ratio of the training instances (default: 0.5)}

\item{colsample_bytree}{Subsample ratio of columns when constructing each tree (default: 0.5)}

\item{objective}{Objective function (default: binary:logistic)}

\item{early_stopping_rounds}{Perform early stopping if no improvement for this many rounds (default: NULL)}

\item{nthread}{Number of parallel threads (default: 1)}

\item{eval_metric}{Evaluation metric (default: error)}

\item{maximize}{Whether to maximize the evaluation metric (default: FALSE)}

\item{feval}{Customized evaluation metric (default: NULL)}

\item{print_every_n}{Print evaluation messages each n-th iterations (default: 1)}

\item{save_period}{Save the model for every given number of rounds (default: NULL)}

\item{save_name}{Name of the saved model file (default: xgboost.model)}

\item{training_set}{File name to output training counts (optional)}

\item{Verbose}{How much details on the progress to print (default: 1)}
}
\description{
Function to train a binary classification model.
}
\examples{
\dontrun{
extdata_path <- system.file("extdata",package = "BagOfMotifs")
motif_counts <- paste0(extdata_path, "/tutorial/Cardiomyocytes_vs_other_counts.txt")


train_binary(input_data = motif_counts, save_name="Cardiomyocytes_vs_other.rds"
, early_stopping_rounds = 100, print_every_n = 100, nthread = 4)
}
}
