% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOM_shap.R
\name{annonTATE}
\alias{annonTATE}
\title{Re-annotates a ggplot created from shapviz. 
Appends gene annotation onto exsiting annotation.}
\usage{
annonTATE(
  plotDat,
  peakAnnotations,
  annotLength = 30,
  waterfallOther = NULL,
  decreasing = FALSE
)
}
\arguments{
\item{plotDat}{Data slot from a ggplot object}

\item{peakAnnotations}{Motif annotation containing columns called "Motif" and "Factor"}

\item{annotLength}{Number of characters to truncate annotation to. Default 30 characters.}

\item{waterfallOther}{Annotation for waterfall plots "other" category}

\item{decreasing}{Plot in decreasing order (default FALSE).}
}
\description{
Re-annotates a ggplot created from shapviz. 
Appends gene annotation onto exsiting annotation.
}
