% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOM_shap.R
\name{annonTATE}
\alias{annonTATE}
\title{Re-annotates a ggplot created from shapviz}
\usage{
annonTATE(plotDat, peakAnnotations, annotLength = 30)
}
\arguments{
\item{plotDat}{Data slot from a ggplot object}

\item{peakAnnotations}{}

\item{annotLength}{Number of characters to truncate annotation to. Default 30 character.}
}
\description{
Re-annotates a ggplot created from shapviz
}
