% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_CREs.R
\name{filterCREs}
\alias{filterCREs}
\title{filterCREs}
\usage{
filterCREs(
  inputBedFile = NULL,
  annotFile = NULL,
  chrSizesFile = NULL,
  u = NULL,
  d = NULL,
  nbp = NULL,
  min_width = 10,
  keep_proximal = FALSE,
  remove_proximal = FALSE,
  non_exonic = FALSE,
  out_bed = NULL,
  inputBedZeroBased = TRUE,
  celloutputDir = NULL,
  minCellPercent = 1,
  ovr_dir = FALSE
)
}
\arguments{
\item{inputBedFile}{input bed filename}

\item{chrSizesFile}{File name of chromsome sizes ( tab 
delimited file with no header: first column chromosome ID, second column size)}

\item{u}{number of basepairs upstream of transcriptional start sites}

\item{d}{number of basepairs downstream of transcriptional start site}

\item{nbp}{number of base pairs}

\item{min_width}{minimum width of a input range to accept (10nt is default)}

\item{keep_proximal}{boolean. Only keep proximal regions (defined by u and d) relative to transcription start sites}

\item{remove_proximal}{Boolean: Remove proximal regions (defined by u and d) relative to transcriptional start sites}

\item{non_exonic}{Boolean - defines if non exonic regions should ONLY be considered.}

\item{out_bed}{filename of output bed file after filtering.}

\item{inputBedZeroBased}{Boolean (default TRUE). Add 1 to start position of input bed file.}

\item{celloutputDir}{directory name (will create) where individual BED files will be populated. If NULL this step is ommited.}

\item{minCellPercent}{Each cell population must make up this percent of all cells (default 1 percent). Will notify what cells are removed}

\item{ovr_dir}{Boolean (default FALSE). Whether to overwrite output directory}
}
\description{
Function to filter a bed file of Cis 
Regulatory Elements (CRE). Will create output bed files.
}
\examples{
\dontrun{
extdata_path <- system.file("extdata",package = "BagOfMotifs")
input.bed <- paste0(extdata_path, "/tutorial/mouseE8.25_peaks.bed")
annot.file <- paste0(extdata_path,"/Mus_musculus.GRCm38.92.gtf.gz")
mouseChrSizes <- paste0(extdata_path,"/mouse_sizes_primary_genome_tab.txt")


 filterCRE(inputBedFile = input.bed, annotFile = annot.file, 
                    remove_proximal = TRUE,  non_exonic = TRUE,
					   u =  1000, d=1000, nbp=500,
                    chrSizesFile = mouseChrSizes, 
                    out_bed =  'mouseE8.25_peaks_filt.bed')
}
}
