% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_binary.R
\name{predict_binary}
\alias{predict_binary}
\title{predict_binary}
\usage{
predict_binary(xgb_model)
}
\arguments{
\item{xgb_model}{File name of model}

\item{motifs}{File containing motif counts}

\item{pred}{File name to output predicted values (default "predictions.txt")}

\item{training_set}{File name to output training counts (optional)}
}
\description{
Function to predict a set of test CREs using a binary classification model
}
\examples{
\dontrun{
extdata_path <- system.file("extdata",package = "BagOfMotifs")
motif_counts <- paste0(extdata_path, "/tutorial/Cardiomyocytes_vs_other_counts.txt")


predict_binary(motifs = motif_counts, xgb_model = "Cardiomyocytes_vs_other.rds")
}
}
