% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_binary.R
\name{predict_binary_multi}
\alias{predict_binary_multi}
\title{predict_binary_multi}
\usage{
predict_binary_multi(
  inputMotif_dir = NULL,
  inputXGB_dir = NULL,
  outputTrain_dir = NULL,
  ncol = 3,
  nrow = 3,
  outputFile = NULL
)
}
\arguments{
\item{motifs}{File containing motif counts}

\item{xgb_model}{File name of model}

\item{pred}{File name to output predicted values (default "predictions.txt")}

\item{training_set}{File name to output training counts (optional)}
}
\description{
Wrapper function for predict_binary. Will run 
on multiple files within specified directory
}
\examples{
\dontrun{
extdata_path <- system.file("extdata",package = "BagOfMotifs")
motif_counts <- paste0(extdata_path, "/tutorial/motifs/Cardiomyocytes_vs_other_counts.txt")


predict_binary(motifs = motif_counts, xgb_model = paste0(extdata_path, "/tutorial/motifs/Cardiomyocytes_vs_other.rds")
}
}
