% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFIMO.R
\name{runFIMO}
\alias{runFIMO}
\title{Function to execute FIMO for at least two sets of CREs specific to different cell types/conditions}
\usage{
runFIMO(
  input_path,
  genome_path,
  motifs_path,
  out_path,
  p_thresh = 1e-04,
  rm_fasta = TRUE
)
}
\arguments{
\item{input_path}{Path to folder containing .bed files with CRE coordinates}

\item{genome_path}{Path to genome file (fasta format)}

\item{motifs_path}{Motif database file (meme format)}

\item{out_path}{Output file folder}

\item{p_thresh}{p-value threshold (default 0.0001)}

\item{rm_fasta}{whether to remove the generated fasta files that contain CRE sequences (default TRUE)}
}
\description{
Function to execute FIMO for at least two sets of CREs specific to different cell types/conditions
}
\examples{

extdata_path <- system.file("extdata",package = "BagOfMotifs")
bed_path <- paste0(extdata_path, "/tutorial/bed_files")


runFIMO(input_path = bed_path, genome_path = "Mus_musculus_GRCm38.fa"
, motifs_path = "gimme.vertebrate.v5.0.meme", out_path = "./fimo_out")
}
