% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOM_shap.R
\name{shapPlots}
\alias{shapPlots}
\title{Plots a waterfall plot of SHAP values for a selected locus}
\usage{
shapPlots(
  shp,
  plotType = "bar",
  annotDat = NULL,
  annotLength = 30,
  order = "decreasing",
  show_numbers = FALSE,
  ...
)
}
\arguments{
\item{shp}{A shapviz object}

\item{plotType}{Plottype, can be "bar" (default), "beeswarm", or "waterfall"}

\item{order}{Ïf set to "decreasing" then plots the highest value first through to lowest value.}

\item{show_numbers}{Boolean. If set to TRUE will display numbers on plot.}

\item{annotData}{Dataframe with columns "Motif" and "Factor". The values in "Motif" should match what was used}

\item{numTF}{Number of transcription factors to plot. The remaining transcriptions factors are plotted under other}
}
\description{
Plots a waterfall plot of SHAP values for a selected locus
}
