% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_CREs.R
\name{textToBED}
\alias{textToBED}
\title{textToBED}
\usage{
textToBED(
  inputTextFile = NULL,
  header = TRUE,
  inputcolnames = c("peak_chr", "peak_start", "peak_end", "celltype_specificity"),
  sep = ",",
  outputFileName = "out.bed",
  removeMultiAnnotatedPeaks = TRUE,
  removeUnnanotatedPeaks = TRUE
)
}
\arguments{
\item{inputTextFile}{Input bed filename}

\item{header}{Boolean (default TRUE). Whether the input file has a header.}

\item{inputcolnames}{Column names corresponding to CRE coordinates and cell type/state annotation. Default to 'c("peak_chr","peak_start", "peak_end", "celltype_specificity")'.}

\item{sep}{Character separating columns. Default to ','.}

\item{outputFileName}{Output file name. Default to "out.bed".}

\item{removeMultiAnnotatedPeaks}{Boolean (default TRUE). Whether to remove CREs annotated to multiple cell types/states.}

\item{removeUnnanotatedPeaks}{Boolean (default TRUE). Whether to remove CREs not annotated to a cell type/state.}
}
\description{
Converts a text file to BED format for peak data
}
\examples{
\dontrun{
extdata_path <- system.file("extdata",package = "BagOfMotifs")
filename <- paste0(extdata_path,'/Pijuan_etal_table_S6.csv.gz')

textToBed(inputTextFile = filename, outputFileName =  'mouseE8.25_peaks_filt.bed')
}

}
